% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss.extra-package.R
\docType{package}
\name{gamlss.extra-package}
\alias{gamlss.extra}
\alias{gamlss.extra-package}
\title{gamlss.extra: Diagnostic tools and sample statistics for the GAMLSS framework}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jaimemosg/gamlss.extra}
  \item Report bugs at \url{https://github.com/Jaimemosg/gamlss.extra/issues}
}

}
\author{
\strong{Maintainer}: Jaime Mosquera \email{jmosquerag@unal.edu.co} (\href{https://orcid.org/0000-0002-1684-4756}{ORCID})

Other contributors:
\itemize{
  \item Freddy Hernandez \email{fhernanb@unal.edu.co} (\href{https://orcid.org/0000-0001-7459-3329}{ORCID}) [contributor]
}

}
\keyword{internal}
